Use tag-multi-select on Assets table
Coding a bit blind here: it's been a while since I worked on this.

There's no way to bind an `input` or `change` event to this input,
since it's a div, so we need to trigger a refresh manually.

Note that tags appear to be non-additive serverside: e.g. if you
select "Artists" and "Curator", you'll get results for both, not
results that are both.